+++ Core/SignalGen/random_signal_gen.m
@@ -66,9 +66,7 @@ classdef random_signal_gen < signal_gen
             end
            
             % domains 
-            
            this.signals_domain = repmat(BreachDomain(), 1, numel(this.signals));
- 
             
         end